Refactored the members view
